requirements.txt
fastapi
uvicorn[standard]
sqlmodel
passlib[argon2,bcrypt]
python-multipart
Pillow
